# ⚙️ feat(issue-229): Add comprehensive encryptionService performance benchmarks